refactor: 경매 presentation layer 관련 내부 DTO를 static inner record로 변경
