--- backend/ddang/src/main/java/com/ddang/ddang/auction/presentation/dto/response/CategoryResponse.java
@@ -1,4 +1,0 @@
-package com.ddang.ddang.auction.presentation.dto.response;
-
-public record CategoryResponse(String main, String sub) {
-}
