--- backend/ddang/src/main/java/com/ddang/ddang/auction/presentation/dto/response/DirectRegionResponse.java
@@ -1,14 +1,0 @@
-package com.ddang.ddang.auction.presentation.dto.response;
-
-import com.ddang.ddang.auction.application.dto.ReadRegionsDto;
-
-public record DirectRegionResponse(String first, String second, String third) {
-
-    public static DirectRegionResponse from(final ReadRegionsDto dto) {
-        return new DirectRegionResponse(
-                dto.firstRegionDto().regionName(),
-                dto.secondRegionDto().regionName(),
-                dto.thirdRegionDto().regionName()
-        );
-    }
-}
