+++ backend/ddang/src/main/java/com/ddang/ddang/auction/presentation/dto/response/ReadAuctionsResponse.java
@@ -1,5 +1,6 @@
 package com.ddang.ddang.auction.presentation.dto.response;
 
+import com.ddang.ddang.auction.application.dto.ReadAuctionDto;
 import com.ddang.ddang.auction.application.dto.ReadAuctionsDto;
 import java.util.List;
 
@@ -16,5 +17,33 @@ public record ReadAuctionsResponse(List<ReadAuctionResponse> auctions, boolean i
                                .toList();
 
         return new ReadAuctionsResponse(readAuctionResponses, readAuctionsDto.isLast());
+    }
+
+    public record ReadAuctionResponse(
+            Long id,
+            String title,
+            String image,
+            int auctionPrice,
+            String status,
+            int auctioneerCount
+    ) {
+        private static ReadAuctionResponse of(final ReadAuctionDto dto, final String imageRelativeUrl) {
+            return new ReadAuctionResponse(
+                    dto.id(),
+                    dto.title(),
+                    imageRelativeUrl + dto.auctionImageStoreNames().get(0),
+                    processAuctionPrice(dto.startPrice(), dto.lastBidPrice()),
+                    dto.auctionStatus().name(),
+                    dto.auctioneerCount()
+            );
+        }
+
+        private static int processAuctionPrice(final Integer startPrice, final Integer lastBidPrice) {
+            if (lastBidPrice == null) {
+                return startPrice;
+            }
+
+            return lastBidPrice;
+        }
     }
 }
